+++ testing/testing-utils/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "testing-utils"
+description = "Testing utils for reth."
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+secp256k1.workspace = true
+alloy-genesis.workspace = true
+reth-primitives.workspace = true
